+++ packages/hoc-input/src/formInput.js
@@ -3,11 +3,9 @@ import PropTypes from "prop-types";
 import classNames from "classnames";
 import wrapDisplayName from "recompose/wrapDisplayName";
 import Text from "@crave/farmblocks-text";
-import Image, { badgeSizes } from "@crave/farmblocks-image";
 import Link from "@crave/farmblocks-link";
 import { fontSizes } from "@crave/farmblocks-theme";
 
-import errorIconSrc from "./constants/errorIcon";
 import Wrapper from "./styledComponents/Wrapper";
 
 export const formInputProps = {
@@ -16,10 +14,8 @@ export const formInputProps = {
   type: PropTypes.string,
   focused: PropTypes.bool,
   disabled: PropTypes.bool,
-  validationMessages: PropTypes.arrayOf(PropTypes.string),
-  errorIconSrc: PropTypes.string,
+  invalid: PropTypes.string,
   onChange: PropTypes.func,
-  onInvalid: PropTypes.func,
   onFocus: PropTypes.func,
   onBlur: PropTypes.func,
   innerRef: PropTypes.func
@@ -31,42 +27,37 @@ const formInput = WrappedComponent => {
       super(props);
       this.state = {
         value: props.value,
-        focused: props.focused,
-        validationMessages: props.validationMessages
+        focused: props.focused
       };
 
       this.onChange = this.onChange.bind(this);
-      this.onInvalid = this.onInvalid.bind(this);
       this.onFocus = this.onFocus.bind(this);
       this.onBlur = this.onBlur.bind(this);
       this._renderInput = this._renderInput.bind(this);
-      this._renderFailedMessages = this._renderFailedMessages.bind(this);
     }
 
     render() {
       const {
         label,
         focused,
-        validationMessages,
-        errorIconSrc,
         onChange,
         onFocus,
         onBlur,
-        onInvalid,
+        invalid,
         ...wrappedComponentProps
       } = this.props;
       const wrapperProps = {
         focused: this.state.focused,
-        invalid: this.state.validationMessages.length > 0,
+        invalid: invalid === "true",
         filled: this.state.value.length > 0,
         disabled: wrappedComponentProps.disabled,
         type: wrappedComponentProps.type
       };
+
       return (
         <Wrapper {...wrapperProps} onClick={this.handleWrapperClick}>
           {this._renderInput(wrappedComponentProps)}
           {this._renderLabel(label)}
-          {this._renderFailedMessages(wrapperProps.invalid)}
         </Wrapper>
       );
     }
@@ -74,7 +65,6 @@ const formInput = WrappedComponent => {
     _renderInput({ innerRef, ...inputProps }) {
       const handlers = {
         onChange: this.onChange,
-        onInvalid: this.onInvalid,
         onFocus: this.onFocus,
         onBlur: this.onBlur
       };
@@ -135,40 +125,12 @@ const formInput = WrappedComponent => {
       );
     }
 
-    _renderFailedMessages(invalid) {
-      return (
-        invalid && (
-          <div className="messages">
-            {this.state.validationMessages.map((text, index) => (
-              <div className="message" key={`err-${index}`}>
-                <Image
-                  className="icon"
-                  src={this.props.errorIconSrc}
-                  badge
-                  size={badgeSizes.SMALL}
-                />
-                <Text
-                  lineHeight={`${badgeSizes.SMALL}px`}
-                  size={fontSizes.MEDIUM}
-                >
-                  {text}
-                </Text>
-              </div>
-            ))}
-          </div>
-        )
-      );
-    }
-
     componentWillReceiveProps(nextProps) {
-      const nextState = {};
       if (nextProps.value !== this.props.value) {
-        nextState.value = nextProps.value;
+        this.setState({
+          value: nextProps.value
+        });
       }
-      if (nextProps.validationMessages !== this.props.validationMessages) {
-        nextState.validationMessages = nextProps.validationMessages;
-      }
-      this.setState(nextState);
     }
 
     handleClearClick = () => {
@@ -188,18 +150,7 @@ const formInput = WrappedComponent => {
       this.setState({
         value: event.value || event.target.value
       });
-      // if no custom validation is provided, clean validation messages once the
-      // browser api considers the value valid
-      if (
-        this.props.validationMessages.length === 0 &&
-        event.target &&
-        event.target.validity &&
-        event.target.validity.valid !== undefined
-      ) {
-        this.setState({
-          validationMessages: []
-        });
-      }
+
       return this.props.onChange(event);
     }
 
@@ -211,20 +162,6 @@ const formInput = WrappedComponent => {
     onBlur(event) {
       this.setState({ focused: false });
       return this.props.onBlur(event);
-    }
-
-    onInvalid(event) {
-      event.preventDefault();
-      // use built-in browser validation message if available and if the validationErrors is not passed
-      if (
-        this.props.validationMessages.length === 0 &&
-        event.target.validationMessage
-      ) {
-        this.setState({
-          validationMessages: [event.target.validationMessage]
-        });
-      }
-      return this.props.onInvalid(event);
     }
 
     static displayName = wrapDisplayName(WrappedComponent, "formInput");
@@ -240,11 +177,8 @@ const formInput = WrappedComponent => {
       focused: false,
       disabled: false,
       onChange: () => null,
-      onInvalid: () => null,
       onFocus: () => null,
-      onBlur: () => null,
-      errorIconSrc,
-      validationMessages: []
+      onBlur: () => null
     };
   };
 };