+++ packages/card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.15"></a>
+## [0.3.15](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-card@0.3.14...@crave/farmblocks-card@0.3.15) (2018-03-15)
+
+
+
+
+**Note:** Version bump only for package @crave/farmblocks-card
+
+<a name="0.3.14"></a>
+## [0.3.14](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-card@0.3.13...@crave/farmblocks-card@0.3.14) (2018-03-14)
+
+
+
+
+**Note:** Version bump only for package @crave/farmblocks-card
+
 <a name="0.3.13"></a>
 ## [0.3.13](https://github.com/CraveFood/farmblocks/compare/@crave/farmblocks-card@0.3.12...@crave/farmblocks-card@0.3.13) (2018-03-01)